+++ src/core/qgsmaptopixelgeometrysimplifier.cpp
@@ -41,7 +41,7 @@ float QgsMapToPixelSimplifier::calculateLengthSquared2D( double x1, double y1, d
 }
 
 //! Returns the BBOX of the specified WKB-point stream
-inline static QgsRectangle calculateBoundingBox( QGis::WkbType wkbType, const unsigned char* wkb, size_t numPoints )
+inline static QgsRectangle calculateBoundingBox( QGis::WkbType wkbType, const unsigned char* wkb, int numPoints )
 {
   double x, y;
   QgsRectangle r;
@@ -50,7 +50,7 @@ inline static QgsRectangle calculateBoundingBox( QGis::WkbType wkbType, const un
   int sizeOfDoubleX = sizeof( double );
   int sizeOfDoubleY = ( QGis::wkbDimensions( wkbType ) - 1 ) * sizeof( double );
 
-  for ( size_t i = 0; i < numPoints; ++i )
+  for ( int i = 0; i < numPoints; ++i )
   {
     memcpy( &x, wkb, sizeof( double ) );
     wkb += sizeOfDoubleX;
@@ -65,9 +65,9 @@ inline static QgsRectangle calculateBoundingBox( QGis::WkbType wkbType, const un
 //! Generalize the WKB-geometry using the BBOX of the original geometry
 static bool generalizeWkbGeometryByBoundingBox(
   QGis::WkbType wkbType,
-  const unsigned char* sourceWkb, size_t sourceWkbSize,
-  unsigned char* targetWkb, size_t& targetWkbSize,
-  const QgsRectangle& envelope, bool writeHeader )
+  const unsigned char *sourceWkb, int sourceWkbSize,
+  unsigned char *targetWkb, int &targetWkbSize,
+  const QgsRectangle &envelope, bool writeHeader )
 {
   Q_UNUSED( sourceWkb );
   unsigned char* wkb2 = targetWkb;
@@ -77,7 +77,7 @@ static bool generalizeWkbGeometryByBoundingBox(
   int sizeOfDoubleY = sizeof( double ) * ( QGis::wkbDimensions( wkbType ) - 1 );
 
   // If the geometry is already minimal skip the generalization
-  size_t minimumSize = geometryType == QGis::WKBLineString ? 4 + 2 * ( sizeOfDoubleX + sizeOfDoubleY ) : 8 + 5 * ( sizeOfDoubleX + sizeOfDoubleY );
+  int minimumSize = geometryType == QGis::WKBLineString ? 4 + 2 * ( sizeOfDoubleX + sizeOfDoubleY ) : 8 + 5 * ( sizeOfDoubleX + sizeOfDoubleY );
 
   if ( writeHeader )
     minimumSize += 5;
@@ -167,9 +167,9 @@ static bool generalizeWkbGeometryByBoundingBox(
 //! Simplify the WKB-geometry using the specified tolerance
 bool QgsMapToPixelSimplifier::simplifyWkbGeometry(
   int simplifyFlags, QGis::WkbType wkbType,
-  const unsigned char* sourceWkb, size_t sourceWkbSize,
-  unsigned char* targetWkb, size_t& targetWkbSize,
-  const QgsRectangle& envelope, double map2pixelTol,
+  const unsigned char *sourceWkb, int sourceWkbSize,
+  unsigned char *targetWkb, int &targetWkbSize,
+  const QgsRectangle &envelope, double map2pixelTol,
   bool writeHeader, bool isaLinearRing )
 {
   bool isGeneralizable = true;
@@ -178,7 +178,7 @@ bool QgsMapToPixelSimplifier::simplifyWkbGeometry(
   // Save initial WKB settings to use when the simplification creates invalid geometries
   const unsigned char* sourcePrevWkb = sourceWkb;
   unsigned char* targetPrevWkb = targetWkb;
-  size_t targetWkbPrevSize = targetWkbSize;
+  int targetWkbPrevSize = targetWkbSize;
 
   // Can replace the geometry by its BBOX ?
   if (( simplifyFlags & QgsMapToPixelSimplifier::SimplifyEnvelope ) &&
@@ -296,7 +296,7 @@ bool QgsMapToPixelSimplifier::simplifyWkbGeometry(
         // approximate the geometry's shape by its bounding box
         // (rect for linear ring / one segment for line string)
         unsigned char* targetTempWkb = targetWkb;
-        size_t targetWkbTempSize = targetWkbSize;
+        int targetWkbTempSize = targetWkbSize;
 
         sourceWkb = sourcePrevWkb;
         targetWkb = targetPrevWkb;
@@ -352,8 +352,8 @@ bool QgsMapToPixelSimplifier::simplifyWkbGeometry(
       memcpy( &numPoints_i, sourceWkb, 4 );
       QgsRectangle envelope_i = numRings == 1 ? envelope : calculateBoundingBox( wkbType, sourceWkb + 4, numPoints_i );
 
-      size_t sourceWkbSize_i = 4 + numPoints_i * QGis::wkbDimensions( wkbType ) * sizeof( double );
-      size_t targetWkbSize_i = 0;
+      int sourceWkbSize_i = 4 + numPoints_i * QGis::wkbDimensions( wkbType ) * sizeof( double );
+      int targetWkbSize_i = 0;
 
       result |= simplifyWkbGeometry( simplifyFlags, wkbType, sourceWkb, sourceWkbSize_i, targetWkb, targetWkbSize_i, envelope_i, map2pixelTol, false, true );
       sourceWkb += sourceWkbSize_i;
@@ -375,8 +375,8 @@ bool QgsMapToPixelSimplifier::simplifyWkbGeometry(
 
     for ( int i = 0; i < numGeoms; ++i )
     {
-      size_t sourceWkbSize_i = 0;
-      size_t targetWkbSize_i = 0;
+      int sourceWkbSize_i = 0;
+      int targetWkbSize_i = 0;
 
       // ... calculate the wkb-size of the current child complex geometry
       if ( flatType == QGis::WKBMultiLineString )
@@ -429,7 +429,7 @@ QgsGeometry* QgsMapToPixelSimplifier::simplify( QgsGeometry* geometry ) const
 {
   QgsGeometry* g = new QgsGeometry();
 
-  size_t wkbSize = geometry->wkbSize();
+  int wkbSize = geometry->wkbSize();
   unsigned char* wkb = reinterpret_cast< unsigned char* >( malloc( wkbSize ) );
   memcpy( wkb, geometry->asWkb(), wkbSize );
   g->fromWkb( wkb, wkbSize );
@@ -441,7 +441,7 @@ QgsGeometry* QgsMapToPixelSimplifier::simplify( QgsGeometry* geometry ) const
 //! Simplifies the geometry (Removing duplicated points) when is applied the specified map2pixel context
 bool QgsMapToPixelSimplifier::simplifyGeometry( QgsGeometry* geometry, int simplifyFlags, double tolerance )
 {
-  size_t finalWkbSize = 0;
+  int finalWkbSize = 0;
 
   // Check whether the geometry can be simplified using the map2pixel context
   QGis::GeometryType geometryType = geometry->type();
@@ -452,7 +452,7 @@ bool QgsMapToPixelSimplifier::simplifyGeometry( QgsGeometry* geometry, int simpl
   QGis::WkbType wkbType = geometry->wkbType();
 
   const unsigned char* wkb = geometry->asWkb();
-  size_t wkbSize = geometry->wkbSize();
+  int wkbSize = geometry->wkbSize();
 
   unsigned char* targetWkb = new unsigned char[wkbSize];
   memcpy( targetWkb, wkb, wkbSize );
